+++ components/sst_importer/src/lib.rs
@@ -24,6 +24,7 @@ pub mod sst_importer;
 
 pub use self::config::Config;
 pub use self::errors::{error_inc, Error, Result};
-pub use self::sst_importer::{sst_meta_to_path, SSTImporter};
+pub use self::import_file::sst_meta_to_path;
+pub use self::sst_importer::SSTImporter;
 pub use self::sst_writer::{RawSSTWriter, TxnSSTWriter};
 pub use self::util::prepare_sst_for_ingestion;
