+++ components/sst_importer/src/sst_importer.rs
@@ -1,7 +1,6 @@
 // Copyright 2018 TiKV Project Authors. Licensed under Apache-2.0.
 
 use std::borrow::Cow;
-use std::collections::HashMap;
 use std::ops::Bound;
 use std::path::{Path, PathBuf};
 use std::sync::Arc;
@@ -12,7 +11,6 @@ use kvproto::backup::StorageBackend;
 use kvproto::import_sstpb::pair::Op as PairOp;
 #[cfg(not(feature = "prost-codec"))]
 use kvproto::import_sstpb::*;
-use uuid::{Builder as UuidBuilder, Uuid};
 
 use encryption::DataKeyManager;
 use engine_rocks::{
@@ -20,14 +18,14 @@ use engine_rocks::{
     RocksSstReader,
 };
 use engine_traits::{
-    name_to_cf, EncryptionKeyManager, IngestExternalFileOptions, Iterator, KvEngine, SSTMetaInfo,
-    SeekKey, SstExt, SstReader, SstWriter, SstWriterBuilder, CF_DEFAULT, CF_WRITE,
+    name_to_cf, EncryptionKeyManager, Iterator, KvEngine, SSTMetaInfo, SeekKey, SstExt, SstReader,
+    SstWriter, SstWriterBuilder, CF_DEFAULT, CF_WRITE,
 };
 use file_system::{get_io_rate_limiter, OpenOptions};
 use tikv_util::time::{Instant, Limiter};
 use txn_types::{Key, TimeStamp, WriteRef};
 
-use crate::import_file::{ImportFile, ImportPath};
+use crate::import_file::{ImportDir, ImportFile};
 use crate::import_mode::{ImportModeSwitcher, RocksDBMetricsFn};
 use crate::metrics::*;
 use crate::sst_writer::{RawSSTWriter, TxnSSTWriter};
@@ -465,198 +463,6 @@ impl SSTImporter {
     }
 }
 
-/// ImportDir is responsible for operating SST files and related path
-/// calculations.
-///
-/// The file being written is stored in `$root/.temp/$file_name`. After writing
-/// is completed, the file is moved to `$root/$file_name`. The file generated
-/// from the ingestion process will be placed in `$root/.clone/$file_name`.
-///
-/// TODO: Add size and rate limit.
-pub struct ImportDir {
-    root_dir: PathBuf,
-    temp_dir: PathBuf,
-    clone_dir: PathBuf,
-}
-
-impl ImportDir {
-    const TEMP_DIR: &'static str = ".temp";
-    const CLONE_DIR: &'static str = ".clone";
-
-    fn new<P: AsRef<Path>>(root: P) -> Result<ImportDir> {
-        let root_dir = root.as_ref().to_owned();
-        let temp_dir = root_dir.join(Self::TEMP_DIR);
-        let clone_dir = root_dir.join(Self::CLONE_DIR);
-        if temp_dir.exists() {
-            file_system::remove_dir_all(&temp_dir)?;
-        }
-        if clone_dir.exists() {
-            file_system::remove_dir_all(&clone_dir)?;
-        }
-        file_system::create_dir_all(&temp_dir)?;
-        file_system::create_dir_all(&clone_dir)?;
-        Ok(ImportDir {
-            root_dir,
-            temp_dir,
-            clone_dir,
-        })
-    }
-
-    fn join(&self, meta: &SstMeta) -> Result<ImportPath> {
-        let file_name = sst_meta_to_path(meta)?;
-        let save_path = self.root_dir.join(&file_name);
-        let temp_path = self.temp_dir.join(&file_name);
-        let clone_path = self.clone_dir.join(&file_name);
-        Ok(ImportPath {
-            save: save_path,
-            temp: temp_path,
-            clone: clone_path,
-        })
-    }
-
-    fn create(
-        &self,
-        meta: &SstMeta,
-        key_manager: Option<Arc<DataKeyManager>>,
-    ) -> Result<ImportFile> {
-        let path = self.join(meta)?;
-        if path.save.exists() {
-            return Err(Error::FileExists(path.save, "create SST upload cache"));
-        }
-        ImportFile::create(meta.clone(), path, key_manager)
-    }
-
-    fn delete_file(&self, path: &Path, key_manager: Option<&DataKeyManager>) -> Result<()> {
-        if path.exists() {
-            file_system::remove_file(&path)?;
-            if let Some(manager) = key_manager {
-                manager.delete_file(path.to_str().unwrap())?;
-            }
-        }
-
-        Ok(())
-    }
-
-    fn delete(&self, meta: &SstMeta, manager: Option<&DataKeyManager>) -> Result<ImportPath> {
-        let path = self.join(meta)?;
-        self.delete_file(&path.save, manager)?;
-        self.delete_file(&path.temp, manager)?;
-        self.delete_file(&path.clone, manager)?;
-        Ok(path)
-    }
-
-    fn exist(&self, meta: &SstMeta) -> Result<bool> {
-        let path = self.join(meta)?;
-        Ok(path.save.exists())
-    }
-
-    fn validate(
-        &self,
-        meta: &SstMeta,
-        key_manager: Option<Arc<DataKeyManager>>,
-    ) -> Result<SSTMetaInfo> {
-        let path = self.join(meta)?;
-        let path_str = path.save.to_str().unwrap();
-        let env = get_encrypted_env(key_manager, None /*base_env*/)?;
-        let env = get_inspected_env(Some(env), get_io_rate_limiter())?;
-        let sst_reader = RocksSstReader::open_with_env(&path_str, Some(env))?;
-        sst_reader.verify_checksum()?;
-        // TODO: check the length and crc32 of ingested file.
-        let meta_info = sst_reader.sst_meta_info(meta.to_owned());
-        Ok(meta_info)
-    }
-
-    fn ingest<E: KvEngine>(
-        &self,
-        metas: &[SstMeta],
-        engine: &E,
-        key_manager: Option<Arc<DataKeyManager>>,
-    ) -> Result<()> {
-        let start = Instant::now();
-
-        let mut paths = HashMap::new();
-        let mut ingest_bytes = 0;
-        for meta in metas {
-            let path = self.join(meta)?;
-            let cf = meta.get_cf_name();
-            super::prepare_sst_for_ingestion(&path.save, &path.clone, key_manager.as_deref())?;
-            ingest_bytes += meta.get_length();
-            engine.reset_global_seq(cf, &path.clone)?;
-            paths.entry(cf).or_insert_with(Vec::new).push(path);
-        }
-
-        let mut opts = E::IngestExternalFileOptions::new();
-        opts.move_files(true);
-        for (cf, cf_paths) in paths {
-            let files: Vec<&str> = cf_paths.iter().map(|p| p.clone.to_str().unwrap()).collect();
-            engine.ingest_external_file_cf(cf, &opts, &files)?;
-        }
-        INPORTER_INGEST_COUNT.observe(metas.len() as _);
-        IMPORTER_INGEST_BYTES.observe(ingest_bytes as _);
-        IMPORTER_INGEST_DURATION
-            .with_label_values(&["ingest"])
-            .observe(start.saturating_elapsed().as_secs_f64());
-        Ok(())
-    }
-
-    fn list_ssts(&self) -> Result<Vec<SstMeta>> {
-        let mut ssts = Vec::new();
-        for e in file_system::read_dir(&self.root_dir)? {
-            let e = e?;
-            if !e.file_type()?.is_file() {
-                continue;
-            }
-            let path = e.path();
-            match path_to_sst_meta(&path) {
-                Ok(sst) => ssts.push(sst),
-                Err(e) => error!(%e; "path_to_sst_meta failed"; "path" => %path.to_str().unwrap(),),
-            }
-        }
-        Ok(ssts)
-    }
-}
-
-const SST_SUFFIX: &str = ".sst";
-
-pub fn sst_meta_to_path(meta: &SstMeta) -> Result<PathBuf> {
-    Ok(PathBuf::from(format!(
-        "{}_{}_{}_{}_{}{}",
-        UuidBuilder::from_slice(meta.get_uuid())?.build(),
-        meta.get_region_id(),
-        meta.get_region_epoch().get_conf_ver(),
-        meta.get_region_epoch().get_version(),
-        meta.get_cf_name(),
-        SST_SUFFIX,
-    )))
-}
-
-fn path_to_sst_meta<P: AsRef<Path>>(path: P) -> Result<SstMeta> {
-    let path = path.as_ref();
-    let file_name = match path.file_name().and_then(|n| n.to_str()) {
-        Some(name) => name,
-        None => return Err(Error::InvalidSSTPath(path.to_owned())),
-    };
-
-    // A valid file name should be in the format:
-    // "{uuid}_{region_id}_{region_epoch.conf_ver}_{region_epoch.version}_{cf}.sst"
-    if !file_name.ends_with(SST_SUFFIX) {
-        return Err(Error::InvalidSSTPath(path.to_owned()));
-    }
-    let elems: Vec<_> = file_name.trim_end_matches(SST_SUFFIX).split('_').collect();
-    if elems.len() != 5 {
-        return Err(Error::InvalidSSTPath(path.to_owned()));
-    }
-
-    let mut meta = SstMeta::default();
-    let uuid = Uuid::parse_str(elems[0])?;
-    meta.set_uuid(uuid.as_bytes().to_vec());
-    meta.set_region_id(elems[1].parse()?);
-    meta.mut_region_epoch().set_conf_ver(elems[2].parse()?);
-    meta.mut_region_epoch().set_version(elems[3].parse()?);
-    meta.set_cf_name(elems[4].to_owned());
-    Ok(meta)
-}
-
 fn key_to_bound(key: &[u8]) -> Bound<&[u8]> {
     if key.is_empty() {
         Bound::Unbounded
@@ -691,26 +497,25 @@ fn is_after_end_bound<K: AsRef<[u8]>>(value: &[u8], bound: &Bound<K>) -> bool {
 
 #[cfg(test)]
 mod tests {
-    use super::*;
-    use test_sst_importer::*;
-
     use std::f64::INFINITY;
-    use tikv_util::stream::block_on_external_io;
+    use std::io;
 
     use engine_traits::{
-        collect, EncryptionMethod, Iterable, Iterator, SeekKey, CF_DEFAULT, DATA_CFS,
+        collect, EncryptionMethod, Error as TraitError, ExternalSstFileInfo, Iterable, Iterator,
+        SeekKey, SstReader, SstWriter, TableProperties, TablePropertiesCollection,
+        TablePropertiesExt, UserCollectedProperties, CF_DEFAULT, DATA_CFS,
     };
-    use engine_traits::{Error as TraitError, TablePropertiesExt};
-    use engine_traits::{
-        ExternalSstFileInfo, TableProperties, TablePropertiesCollection, UserCollectedProperties,
-    };
+    use file_system::File;
     use tempfile::Builder;
-    use test_sst_importer::{
-        new_sst_reader, new_sst_writer, new_test_engine, RocksSstWriter, PROP_TEST_MARKER_CF_NAME,
-    };
-    use txn_types::{Value, WriteType};
-
+    use test_sst_importer::*;
     use test_util::new_test_key_manager;
+    use tikv_util::stream::block_on_external_io;
+    use txn_types::{Value, WriteType};
+    use uuid::Uuid;
+
+    use super::*;
+    use crate::import_file::ImportPath;
+    use crate::*;
 
     fn do_test_import_dir(key_manager: Option<Arc<DataKeyManager>>) {
         let temp_dir = Builder::new().prefix("test_import_dir").tempdir().unwrap();
@@ -870,24 +675,6 @@ mod tests {
         // test with tde
         let (_tmp_dir, key_manager) = new_key_manager_for_test();
         do_test_import_file(Some(key_manager));
-    }
-
-    #[test]
-    fn test_sst_meta_to_path() {
-        let mut meta = SstMeta::default();
-        let uuid = Uuid::new_v4();
-        meta.set_uuid(uuid.as_bytes().to_vec());
-        meta.set_region_id(1);
-        meta.set_cf_name(CF_DEFAULT.to_owned());
-        meta.mut_region_epoch().set_conf_ver(2);
-        meta.mut_region_epoch().set_version(3);
-
-        let path = sst_meta_to_path(&meta).unwrap();
-        let expected_path = format!("{}_1_2_3_default.sst", uuid);
-        assert_eq!(path.to_str().unwrap(), &expected_path);
-
-        let new_meta = path_to_sst_meta(path).unwrap();
-        assert_eq!(meta, new_meta);
     }
 
     fn create_external_sst_file_with_write_fn<F>(
