+++ tests/integrations/config/dynamic/raftstore.rs
@@ -76,7 +76,7 @@ fn start_raftstore(
             .as_path()
             .display()
             .to_string();
-        Arc::new(SSTImporter::new(&cfg.import, &p, None).unwrap())
+        Arc::new(SSTImporter::new(&cfg.import, &p, None, false).unwrap())
     };
     let snap_mgr = {
         let p = dir
